issue-1038: dataset.Types added
